+++ packages/riverpod_lint/CHANGELOG.md
@@ -3,6 +3,8 @@
 - Disable unsupported_provider_value when a notifier returns "this"
 - Fix scoped_providers_should_specify_dependencies incorrectly triggering on functions other than "main"
 - Handle cascade operators in ref expressions
+- Fix `provider_dependencies` not considering dependencies inside methods
+  other than `build` of a notifier.
 
 ## 1.1.7 - 2023-04-06